Nanominer Exiting unexpectedly!
My nanominer version 3.3.5 is failing to update to version 3.3.6 and showing the message in green color as "Added option 'maxRejectedShares': set the maximum amount of rejected shares before restarting miner process...etc. Downloading the new version (will restart when it is downloaded)." After this the miner runs for few seconds and then it is exiting from the screen and is not updating as before by downloading the new version 3.3.6. Please help me to solve this issue as I do no understand which line of code i shall use in the config.ini file so that miner shall work properly.